seperated the classes
