+++ app/src/main/java/edu/uc/hornunjb/weatherwear/dto/Clouds.kt
@@ -1,0 +1,8 @@
+package edu.uc.hornunjb.weatherwear.dto
+
+import com.google.gson.annotations.SerializedName
+
+class Clouds (
+    @SerializedName("all")
+    var all: Float = 0.toFloat()
+)
