+++ app/src/main/java/edu/uc/hornunjb/weatherwear/dto/Main.kt
@@ -1,0 +1,16 @@
+package edu.uc.hornunjb.weatherwear.dto
+
+import com.google.gson.annotations.SerializedName
+
+class Main (
+    @SerializedName("temp")
+    var temp: Float = 0.toFloat(),
+    @SerializedName("humidity")
+    var humidity: Float = 0.toFloat(),
+    @SerializedName("pressure")
+    var pressure: Float = 0.toFloat(),
+    @SerializedName("temp_min")
+    var temp_min: Float = 0.toFloat(),
+    @SerializedName("temp_max")
+    var temp_max: Float = 0.toFloat()
+)
