+++ app/src/main/java/edu/uc/hornunjb/weatherwear/dto/Rain.kt
@@ -1,0 +1,8 @@
+package edu.uc.hornunjb.weatherwear.dto
+
+import com.google.gson.annotations.SerializedName
+
+class Rain (
+    @SerializedName("3h")
+    var h3: Float = 0.toFloat()
+)
