+++ app/src/main/java/edu/uc/hornunjb/weatherwear/dto/Sys.kt
@@ -1,0 +1,12 @@
+package edu.uc.hornunjb.weatherwear.dto
+
+import com.google.gson.annotations.SerializedName
+
+class Sys (
+    @SerializedName("country")
+    var country: String? = null,
+    @SerializedName("sunrise")
+    var sunrise: Long = 0,
+    @SerializedName("sunset")
+    var sunset: Long = 0
+)
