+++ app/src/main/java/edu/uc/hornunjb/weatherwear/dto/Weather.kt
@@ -3,31 +3,7 @@ package edu.uc.hornunjb.weatherwear.dto
 import com.google.gson.annotations.SerializedName
 import java.util.ArrayList
 
-class WeatherResponse {
 
-    @SerializedName("coord")
-    var coord: Coord? = null
-    @SerializedName("sys")
-    var sys: Sys? = null
-    @SerializedName("weather")
-    var weather = ArrayList<Weather>()
-    @SerializedName("main")
-    var main: Main? = null
-    @SerializedName("wind")
-    var wind: Wind? = null
-    @SerializedName("rain")
-    var rain: Rain? = null
-    @SerializedName("clouds")
-    var clouds: Clouds? = null
-    @SerializedName("dt")
-    var dt: Float = 0.toFloat()
-    @SerializedName("id")
-    var id: Int = 0
-    @SerializedName("name")
-    var name: String? = null
-    @SerializedName("cod")
-    var cod: Float = 0.toFloat()
-}
 
 class Weather {
     @SerializedName("id")
@@ -40,44 +16,15 @@ class Weather {
     var icon: String? = null
 }
 
-class Clouds {
-    @SerializedName("all")
-    var all: Float = 0.toFloat()
-}
 
-class Rain {
-    @SerializedName("3h")
-    var h3: Float = 0.toFloat()
-}
 
-class Wind {
-    @SerializedName("speed")
-    var speed: Float = 0.toFloat()
-    @SerializedName("deg")
-    var deg: Float = 0.toFloat()
-}
 
-class Main {
-    @SerializedName("temp")
-    var temp: Float = 0.toFloat()
-    @SerializedName("humidity")
-    var humidity: Float = 0.toFloat()
-    @SerializedName("pressure")
-    var pressure: Float = 0.toFloat()
-    @SerializedName("temp_min")
-    var temp_min: Float = 0.toFloat()
-    @SerializedName("temp_max")
-    var temp_max: Float = 0.toFloat()
-}
 
-class Sys {
-    @SerializedName("country")
-    var country: String? = null
-    @SerializedName("sunrise")
-    var sunrise: Long = 0
-    @SerializedName("sunset")
-    var sunset: Long = 0
-}
+
+
+
+
+
 
 class Coord {
     @SerializedName("lon")
