+++ app/src/main/java/edu/uc/hornunjb/weatherwear/dto/Wind.kt
@@ -1,0 +1,10 @@
+package edu.uc.hornunjb.weatherwear.dto
+
+import com.google.gson.annotations.SerializedName
+
+class Wind (
+    @SerializedName("speed")
+    var speed: Float = 0.toFloat(),
+    @SerializedName("deg")
+    var deg: Float = 0.toFloat()
+    )
