+++ app/src/main/java/edu/uc/hornunjb/weatherwear/ui/main/MainFragment.kt
@@ -24,7 +24,7 @@ class MainFragment : Fragment() {
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
         viewModel = ViewModelProviders.of(this).get(MainViewModel::class.java)
-        // TODO: Use the ViewModel
+
     }
 
 }
