+++ app/src/main/java/edu/uc/hornunjb/weatherwear/ui/main/MainViewModel.kt
@@ -1,7 +1,10 @@
 package edu.uc.hornunjb.weatherwear.ui.main
 
+import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import edu.uc.hornunjb.weatherwear.dto.WeatherResponse
 
 class MainViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
+
+
 }
