+++ local.properties
@@ -4,5 +4,5 @@
 # Location of the SDK. This is only used by Gradle.
 # For customization when using a Version Control System, please read the
 # header note.
-#Thu Mar 04 11:47:48 EST 2021
-sdk.dir=C\:\\Users\\jacob\\AppData\\Local\\Android\\Sdk
+#Sun Mar 07 14:42:52 EST 2021
+sdk.dir=C\:\\Users\\meyer\\AppData\\Local\\Android\\Sdk
